requirements.txt
# Core dependencies
telethon==1.24.0
pandas==1.5.3
sqlalchemy==2.0.10
alembic==1.10.3

# FastAPI and Web
fastapi==0.95.1
uvicorn[standard]==0.21.1
pydantic==1.10.7
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.1.0
aiohttp==3.8.4

# Database
sqlite3
psycopg2-binary==2.9.6  # PostgreSQL support

# YOLO and Computer Vision
ultralytics==8.0.100
opencv-python==4.7.0.72
Pillow==9.5.0
torch>=1.13.0
torchvision>=0.14.0

# Natural Language Processing
nltk==3.8.1
spacy==3.5.2
textblob==0.17.1
langdetect==1.0.9

# Data Analysis and Visualization
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.10.1
scikit-learn==1.2.2
plotly==5.14.1

# Async and Concurrency
aiofiles==23.1.0
aiohttp==3.8.4
asyncio-throttle==1.0.2

# Utilities
requests==2.28.2
python-dotenv==1.0.0
click==8.1.3
tqdm==4.65.0
loguru==0.7.0
pytz==2023.3
dateutil==2.8.2
validators==0.20.0

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.1.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0
httpx==0.24.1  # For testing FastAPI

# Development
black==23.3.0
flake8==6.0.0
isort==5.12.0
pre-commit==3.3.2
mypy==1.3.0

# Monitoring and Logging
prometheus-client==0.16.0
structlog==23.1.0

# Security
cryptography==40.0.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Optional: Dagster (for advanced pipeline orchestration)
# dagster==1.3.3
# dagster-webserver==1.3.3

# Optional: dbt (for data transformation)
# dbt-core==1.4.5
# dbt-sqlite==1.4.0